codebase-query-contextualiser/function-calling/codebase/requirement.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-openai
langchain-community
faiss-cpu
python-dotenv
